requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.2.0
pytesseract==0.3.10
pdf2image==1.17.0
